Updated the library managemnt's functions
