+++ storage/.gitignore
@@ -2,3 +2,4 @@ __pycache__/
 
 *.xlsx
 *.xls
+*.env
